webapi/webapi-messaging-w3c-tests/messaging/SmsManager.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Shentu.Jiazhen <jiazhenx.shentu@intel.com>
        Jianfeng.Xu <jianfengx.xu@intel.com>

-->

<meta charset='utf-8'>
<title>Messaging Test: SmsManager</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/2012/sysapps/messaging/#smsmanager-interface">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="support/support.js"></script>
<div id="log"></div>
<script>

var t1 = async_test("Check if the segments of SmsSegmentInfo is the total number of SMS segments needed to send the input string", {timeout: 2000});
var t2 = async_test("Check if the charsPerSegment of SmsSegmentInfo is the number of characters available per SMS segment as per the encoding to be used to send the SMS message", {timeout: 2000});
var t3 = async_test("Check if the charsAvailableInLastSegment of SmsSegmentInfo is the maximum number of available characters in the last segment needed to send the input string", {timeout: 2000});
var t4 = async_test("Check if the clear method of SmsManager delete all the messages associated to the messaging service", {timeout: 2000});

t1.step(function () {
  assert_true(!!message, "The define of message");
  message.sms.segmentInfo("test").then(function (data) {
    t1.step(function () {
      assert_true(!!data.segments, "The segments of smsSegmentInfo");
    });
    t1.done();
  }, function (error) {
    t1.step(function () {
      assert_unreached(error.message);
    });
    t1.done();
  });
});

t2.step(function () {
  assert_true(!!message, "The define of message");
  message.sms.segmentInfo("test").then(function (data) {
    t2.step(function () {
      assert_true(!!data.charsPerSegment, "The charsPerSegment of smsSegmentInfo");
    });
    t2.done();
  }, function (error) {
    t2.step(function () {
      assert_unreached(error.message);
    });
    t2.done();
  });
});

t3.step(function () {
  assert_true(!!message, "The define of message");
  message.sms.segmentInfo("test").then(function (data) {
    t3.step(function () {
      assert_true(!!data.charsAvailableInLastSegment, "The charsAvailableInLastSegment of smsSegmentInfo");
    });
    t3.done();
  }, function (error) {
    t3.step(function () {
      assert_unreached(error.message);
    });
    t3.done();
  });
});

t4.step(function () {
  assert_true(!!message, "The define of message");
  for (var serviceID in message.sms.serviceIDs )
  {
    message.sms.clear(serviceID).then(function (data) {
      t4.step(function () {
        assert_true(!!data, "A object will be returned in order to notify the result of the request");
      });
    t4.done();
    }, function (error) {
      t4.step(function () {
        assert_unreached(error.message);
      });
      t4.done();
    });
  }
});

test(function () {
  assert_true(!!message, "The define of message");
  assert_equals(message.sms.type, "sms", "The type of SmsManager");
}, "Check if the type attribute is the type of the messaging service manager");

test(function () {
  assert_true(!!message, "The define of message");
  assert_equals(message.sms.serviceIDs[0], "sim0", "The serviceIDs of SmsManager");
}, "Check if the serviceIDs attribute is the identifier of the different services for this type of messaging service");
</script>
